refact: リファクタリング
refact: NotionApiClientをNotionApiServiceに統合

refact: 以前notionのmodelとして実装した不要なdata classを削除

refact: 使ってないgsonライブラリをgradleから削除
